Refactored API tests via TestWiresAPIMixin.
