+++ tests/test_instance.py
@@ -20,174 +20,6 @@ from . import helpers
 
 
 
-class TestWiresAPI(unittest.TestCase):
-
-    """
-    Minimal API tests for the Wires singleton.
-    """
-
-    def setUp(self):
-
-        self.w = Wiring()
-
-
-    def test_unwired_call_does_not_fail(self):
-        """
-        Calling an unwired call works. Does nothing, but works.
-        """
-        self.w.unwired_call()
-
-
-    def test_wiring_non_callable_raises_value_error(self):
-        """
-        Wiring a call to a non-callable raises ValueError.
-        The exception argument (message):
-        - Starts with "argument not callable: ".
-        - Contains repr(argument).
-        """
-        for non_callable in (None, True, 42, 2.3, (), [], {}, set()):
-
-            with self.assertRaises(ValueError) as cm:
-                self.w.wire.this.calls_to(non_callable)
-
-            exception_args = cm.exception.args
-            self.assertEqual(len(exception_args), 1)
-
-            msg = exception_args[0]
-            self.assertTrue(
-                msg.startswith('argument not callable: '),
-                'wrong exception message: %r' % (msg,),
-            )
-            self.assertIn(
-                repr(non_callable),
-                msg,
-                'missing argument repr in message: %r' % (msg,),
-            )
-
-
-    @staticmethod
-    def _test_callable():
-        pass
-
-
-    def test_wiring_callable_works(self):
-        """
-        Wiring a callable works.
-        """
-        # Calling test_callble works (for coverage completion's sake).
-        self._test_callable()
-
-        self.w.wire.this.calls_to(self._test_callable)
-        self.addCleanup(self._unwire_test_callable)
-
-
-    def _unwire_test_callable(self):
-
-        self.w.unwire.this.calls_to(self._test_callable)
-
-
-    def test_wiring_unwiring_works(self):
-        """
-        Wiring and then unwiring to the same callable works.
-        """
-        # Calling test_callble works (for coverage completion's sake).
-        self._test_callable()
-
-        self.w.wire.this.calls_to(self._test_callable)
-        self.w.unwire.this.calls_to(self._test_callable)
-
-
-    def test_unwiring_unknown_callable_raises_value_error(self):
-        """
-        Unwiring an unknown callable raises a ValueError.
-        The exception argument (message):
-        - Starts with "unknown function ".
-        """
-        with self.assertRaises(ValueError) as cm:
-            self.w.unwire.this.calls_to(self._test_callable)
-
-        exception_args = cm.exception.args
-        self.assertEqual(len(exception_args), 1)
-
-        msg = exception_args[0]
-        self.assertTrue(
-            msg.startswith('unknown function '),
-            'wrong exception message: %r' % (msg,),
-        )
-
-
-    def test_dynamic_name_wire_call_unwire(self):
-        """
-        Wiring/unwiring via indexing works.
-        """
-        name = 'name'
-        self.w.wire[name].calls_to(self._test_callable)
-        self.w.unwire[name].calls_to(self._test_callable)
-
-
-    def test_call_via_wire_fails(self):
-        """
-        """
-        with self.assertRaises(RuntimeError) as cm:
-            self.w.wire.some_callable()
-
-        # TODO: assert exception details
-
-
-    def test_call_via_unwire_fails(self):
-        """
-        """
-        with self.assertRaises(RuntimeError) as cm:
-            self.w.unwire.some_callable()
-
-        # TODO: assert exception details
-
-
-    def test_callable_get_use_log_via_wire_fails(self):
-        """
-        """
-        with self.assertRaises(RuntimeError) as cm:
-            _ = self.w.wire.some_callable.use_log
-
-        # TODO: assert exception details
-
-
-    def test_callable_get_use_log_via_unwire_fails(self):
-        """
-        """
-        with self.assertRaises(RuntimeError) as cm:
-            _ = self.w.unwire.some_callable.use_log
-
-        # TODO: assert exception details
-
-
-    def test_callable_set_use_log_via_wire_fails(self):
-        """
-        """
-        with self.assertRaises(RuntimeError) as cm:
-            self.w.wire.some_callable.use_log = None
-
-        # TODO: assert exception details
-
-
-    def test_callable_set_use_log_via_unwire_fails(self):
-        """
-        """
-        with self.assertRaises(RuntimeError) as cm:
-            self.w.unwire.some_callable.use_log = None
-
-        # TODO: assert exception details
-
-
-    def test_wiring_from_instance_fails(self):
-
-        with self.assertRaises(RuntimeError) as cm:
-            self.w.some_callable.calls_to(self._test_callable)
-
-        # TODO: assert exception details
-
-
-
 class TestWiresUtilization(helpers.CallTrackerAssertMixin, unittest.TestCase):
 
     """
@@ -622,7 +454,5 @@ class TestWiresDoubleFullPassing(ArgPassingMixin, unittest.TestCase):
     expected_2nd_call_args = (4, 5, 6, 7, 8, 9)
     expected_2nd_call_kwargs = dict(c='c', d='d', e='e', f='f')
 
-
-# ----------------------------------------------------------------------------
 
 # ----------------------------------------------------------------------------
